feat (api): mv scheduler package to queue package
